// src/app/core/models/settings/radiologist-signup/HrSubspecialty.ts
export class HrSubspecialty {
  SubspecialtyId!: number;
  SubspecialtyUid!: string;
  SpecialtyId?: number;
  SubspecialtyName!: string;
  SubspecialtyDetail!: string;
  IsActive?: boolean;
  IsUpdated?: boolean;
  IsDeleted?: boolean;
  OrgId?: number;
  CreatedBy?: number;
  CreatedOn?: Date;
  LastModifiedBy?: number;
  LastModifiedOn?: Date;
}